Fixed subchapter 1 challenge 1

+++ section_2/2-1/main.py
@@ -1,6 +1,6 @@
 # Required variables
 PI = 3.14159
-radius = int(input())
+radius = float(input())
 
 # Create the function to calculate the area of a circle
 def calculate_circle_area(PI, radius):
